+++ spec/circleci/parallel/task/mock_slave_spec.rb
@@ -13,7 +13,7 @@ module CircleCI::Parallel
     end
 
     let(:configuration) do
-      Configuration.new
+      Configuration.new.slave_node_configuration
     end
 
     it 'creates join marker file' do